rename get and set pointer, update license, bump version
