+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@jdw/jst",
-  "version": "2.0.0-beta.2",
+  "version": "2.0.0-beta.3",
   "description": "A json schema toolkit.",
   "keywords": [
     "json-schema"
