+++ src/get.ts
@@ -1,7 +1,37 @@
+// The `get` function dereferences [json
+// pointers](https://tools.ietf.org/html/rfc6901) according to the IETF RFC6901
+// specification. It takes a `schema` and a json `pointer` as its arguments and
+// then returns the value in `schema` described by `pointer` throwing an error
+// if the path described by the `pointer` was not found in the `schema`.
+//
+// ## Usage:
+//
+// ```
+// import { get } from '@jdw/jst';
+//
+// const schema = { definitions: { foo: 123 }};
+// get(schema, '#/definitions/foo'); // resolves value 123
+// ```
+//
+// **Arguments**
+// - `schema` {Object}: The object to query.
+// - `pointer` {string}: The JSON pointer path of the property to get.
+//
+// **Returns**
+// - {any}: The resolved pointer value.
+//
+// **Throws**
+// - {Error}: If `pointer` cannot be resolved in `schema`.
+//
+// ## Dependencies
+
 import { has } from 'lodash';
 
-const setPointer: Jst.setPointer = (obj, pointer, value) => {
+// #### Implementation
 
+// The `get` function implements the `Jst.Getter` signature as described in
+// [typings.d.ts](typings.d.html).
+const get: Jst.getPointer = (schema, pointer) => {
   // A JSON `pointer` must begin with the symbols '#', '/' or be an empty
   // string ''. So as a first step, we check that this assumption is true and
   // bail if not.
@@ -14,53 +44,46 @@ const setPointer: Jst.setPointer = (obj, pointer, value) => {
   // iterate `schema` checking from left to right that each piece of `pointer`
   // references a valid path in `schema`.
   const fragments = pointer.split('/');
-  let ref = obj;
-  return fragments.forEach((fragment, index) => {
+  return fragments.reduce((object, fragment) => {
     // If `fragment` points to the root path of `object` we can just return
     // the `object` itself.
     if (fragment === '#' || fragment === '/' || fragment === '') {
-      return;
+      return object;
     }
 
     // Here we decode `fragment` according to the JSON pointer
     // specification, replacing the character codes '~1' and '~0' with the
     // symbols '/' and '~' respectively.
     const token = fragment.replace('~1', '/').replace('~0', '~');
+    let reference = null;
 
     // If the 'object' is array assume that `token` indicates an index in
     // this array and try to resolve it appropriately.
-    if (Array.isArray(ref)) {
-      const i = parseInt(token, 10);
+    if (Array.isArray(object)) {
+      const index = parseInt(token, 10);
 
-      if (!ref.indexOf(i)) {
+      if (!object.indexOf(index)) {
         throw new Error(
           `could not dereference JSON pointer: ${pointer}. Array does not have`
-          + ` index: ${index}::${JSON.stringify(obj)}`);
+          + ` index: ${index}::${JSON.stringify(object)}`);
       }
 
-      if ((index + 1) === fragments.length) {
-        ref[i] = value;
-        return;
-      }
-
-      ref = ref[i];
+      reference = object[index];
     }
     // Otherwise if `object` *is not* an Array we expect `object` to be of
     // type Object and that `token` references a valid path in `object`.
     else {
-      if (!has(ref, token)) {
+      if (!has(object, token)) {
         throw new Error(
           `could not dereference pointer '${pointer}'. The fragment ${token}`
-          + ` is not a valid property of object: ${JSON.stringify(obj, null, 2)}`);
+          + ` is not a valid property of object: ${JSON.stringify(object, null, 2)}`);
       }
-
-      if ((index + 1) === fragments.length) {
-        ref[token] = value;
-        return;
-      }
-      ref = ref[token];
+      reference = object[token];
     }
-  });
+
+    // Now return `reference`
+    return reference;
+  }, schema);
 };
 
-export default setPointer;
+export default get;
